Edit general setting validation

+++ app/Helper/helpers.php
@@ -4,43 +4,45 @@
 
 /** Handle file upload */
 
-function handleUpload($inputName, $model=null){
-    try{
-        if(request()->hasFile($inputName)){
-            if($model && \File::exists(public_path($model->{$inputName}))) {
+function handleUpload($inputName, $model = null)
+{
+    try {
+        if (request()->hasFile($inputName)) {
+            if ($model && \File::exists(public_path($model->{$inputName})) && $model->{$inputName} != 'default/logo.png') {
                 \File::delete(public_path($model->{$inputName}));
             }
 
             $file = request()->file($inputName);
-            $fileName = rand().$file->getClientOriginalName();
+            $fileName = rand() . $file->getClientOriginalName();
             $file->move(public_path('/uploads'), $fileName);
 
-            $filePath = "/uploads/".$fileName;
+            $filePath = "/uploads/$fileName";
 
             return $filePath;
         }
-    }catch(\Exception $e){
+    } catch (\Exception $e) {
         throw $e;
     }
-
 }
 
 
 /** Delete file */
 
-function deleteFileIfExist($filePath){
-    try{
-        if(\File::exists(public_path($filePath))){
+function deleteFileIfExist($filePath)
+{
+    try {
+        if (\File::exists(public_path($filePath))) {
             \File::delete(public_path($filePath));
         }
-    }catch(\Exception $e){
+    } catch (\Exception $e) {
         throw $e;
     }
 }
 
 /** get dynamic colors  */
 
-function getColor($index){
+function getColor($index)
+{
     $colors = ['#558bff', '#fecc90', '#ff885e', '#282828', '#190844', '#9dd3ff'];
 
     return $colors[$index % count($colors)];
@@ -48,10 +50,11 @@ function getColor($index){
 
 /** Set Sidebar Active  */
 
-function setSidebarActive($route){
-    if(is_array($route)){
-        foreach($route as $r){
-            if(request()->routeIs($r)){
+function setSidebarActive($route)
+{
+    if (is_array($route)) {
+        foreach ($route as $r) {
+            if (request()->routeIs($r)) {
                 return 'active';
             }
         }
